requirements.txt
annotated-types==0.6.0
asgiref==3.7.2
certifi==2024.2.2
cffi==1.16.0
charset-normalizer==3.3.2
Django==5.0.2
django-jazzmin==2.6.0
django-ninja==1.1.0
gunicorn==21.2.0
idna==3.6
packaging==23.2
passlib==1.7.4
pycparser==2.21
pydantic==2.6.1
pydantic_core==2.16.2
PyMySQL==1.1.0
PyNaCl==1.5.0
requests==2.31.0
sqlparse==0.4.4
typing_extensions==4.9.0
urllib3==2.2.0
